backend/requirements.txt
tensorflow
transformers
opencv-python
flask
ffmpeg-python
